// shared/src/types.ts
export interface IExperience {
  _id: string;
  company: string;
  position: string;
  location: string;
  startDate: Date;
  endDate?: Date;
  current: boolean;
  description: string;
  responsibilities: string[];
  achievements: string[];
  technologies: string[];
  order: number;
  visibility: boolean;
  createdAt: Date;
  updatedAt: Date;
  createdBy: string;
}